Optimized federal tax calculation view.
Simplified the FederalTaxCalculateView by modularizing the tax calculation logic into separate methods. Introduced dynamic tax data retrieval with a fallback to 2024 data.
