+++ src/pages/views/state_calculator_views.py
@@ -1,0 +1,5 @@
+from django.views.generic import TemplateView
+
+
+class StateCalculatorView(TemplateView):
+    template_name = 'pages/state-calculator.html'
